melhorado o controller com novas tratativas

+++ src/controllers/wallets.controller.ts
@@ -40,8 +40,13 @@ class WalletsController {
 
 	private deleteWallet = async (req: Request, res: Response) => {
 		const { id } = req.params;
+
 		try {
-			await walletsService.deleteWallet(Number(id));
+			const wallet = await walletsService.findWalletById(Number(id))
+			if (!wallet) {
+				res.status(404).json({ message: "Wallet not found" });
+				return;
+			}
 			res.status(200).json({ message: "Wallet deleted successfully" });
 		} catch (error) {
 			this.handleError(res, error, "Error deleting wallet");
@@ -52,43 +57,66 @@ class WalletsController {
 		const { id } = req.params;
 		const { newName } = req.body;
 		if (!newName) {
-			res.status(400).json({ message: "Missing newName in request body" });
+			res.status(400).json({ message: "Missing new name in request body" });
 			return;
 		}
 
 		try {
+			const existingWallet = await walletsService.findWalletById(Number(id));
+			if (!existingWallet) {
+				res.status(404).json({ message: "Wallet not found" });
+				return;
+			}
 			const wallet = await walletsService.updateWalletName(Number(id), newName);
 			res.status(200).json({ message: "Wallet name updated", data: wallet });
 		} catch (error) {
 			this.handleError(res, error, "Error updating wallet name");
 		}
-	}
+	};
 
 	private addUserToWallet = async (req: Request, res: Response) => {
 		const { walletId } = req.params;
 		const { userId } = req.body;
 		if (!userId) {
-			res.status(400).json({ message: "Missing userId in request body" });
+			res.status(400).json({ message: "Missing user ID in request body" });
 			return;
 		}
 
 		try {
+			const existingWallet = await walletsService.findWalletById(Number(walletId));
+			if (!existingWallet) {
+				res.status(404).json({ message: "Wallet not found" });
+				return;
+			}
+
 			const wallet = await walletsService.addUserToWallet(Number(walletId), Number(userId));
 			res.status(200).json({ message: "User added to wallet", data: wallet });
 		} catch (error) {
 			this.handleError(res, error, "Error adding user to wallet");
 		}
-	}
+	};
 
 	private removeUserFromWallet = async (req: Request, res: Response) => {
 		const { walletId, userId } = req.params;
 		try {
-			const wallet = await walletsService.removeUserFromWallet(Number(walletId), Number(userId));
+			const existingWallet = await walletsService.findWalletById(Number(walletId));
+			if (!existingWallet) {
+				res.status(404).json({ message: "Wallet not found" });
+				return;
+			}
+
+			const existingUser = await walletsService.findUserInWallet(Number(walletId), Number(userId))
+			if (!existingUser) {
+				res.status(404).json({ message: "User not found in wallet" });
+				return;
+			}
+
+			const wallet = await walletsService.removeUserFromWallet(Number(walletId), Number(userId))
 			res.status(200).json({ message: "User removed from wallet", data: wallet });
 		} catch (error) {
 			this.handleError(res, error, "Error removing user from wallet");
 		}
-	}
+	};
 
 	private getWallets = async (_req: Request, res: Response) => {
 		try {
@@ -116,12 +144,18 @@ class WalletsController {
 	private getWalletUsers = async (req: Request, res: Response) => {
 		const { id } = req.params;
 		try {
+			const existingWallet = await walletsService.findWalletById(Number(id));
+			if (!existingWallet) {
+				res.status(404).json({ message: "Wallet not found" });
+				return;
+			}
+
 			const userIds = await walletsService.getWalletUsers(Number(id));
 			res.status(200).json({ data: userIds });
 		} catch (error) {
 			this.handleError(res, error, "Error fetching wallet users");
 		}
-	}
+	};
 
 	private getUserWallets = async (req: Request, res: Response) => {
 		const { userId } = req.params;
@@ -133,10 +167,7 @@ class WalletsController {
 		}
 
 		try {
-			const wallets = await walletsService.getUserWallets(
-				instance as string,
-				Number(userId)
-			);
+			const wallets = await walletsService.getUserWallets(String(instance), Number(userId))
 			res.status(200).json({ data: wallets });
 		} catch (error) {
 			this.handleError(res, error, "Error fetching user wallets");
@@ -146,17 +177,18 @@ class WalletsController {
 	private findUserInWallet = async (req: Request, res: Response) => {
 		const { walletId, userId } = req.params;
 		try {
-			const relation = await walletsService.findUserInWallet(
-				Number(walletId),
-				Number(userId)
-			);
-
-			if (!relation) {
+			const existingWallet = await walletsService.findWalletById(Number(walletId));
+			if (!existingWallet) {
+				res.status(404).json({ message: "Wallet not found" });
+				return;
+			}
+			const user = await walletsService.findUserInWallet(Number(walletId), Number(userId))
+			if (!user) {
 				res.status(404).json({ message: "User not found in wallet" });
 				return;
 			}
 
-			res.status(200).json({ data: relation });
+			res.status(200).json({ data: user });
 		} catch (error) {
 			this.handleError(res, error, "Error finding user in wallet");
 		}
